+++ examples/technical_details/plot_diagnostic_api.py
@@ -41,12 +41,15 @@ deep_tree = DecisionTreeRegressor(random_state=42)
 
 # %%
 # Calling :meth:`~skore.EstimatorReport.diagnosis.summarize` explicitly
-# ===========================================================
+# =====================================================================
 #
-# Every report exposes a :meth:`~skore.EstimatorReport.diagnosis` accessor.
-# Checks are computed lazily and cached, so calling
-# :meth:`~skore.EstimatorReport.diagnosis.summarize` is always cheap after the first
-# call.
+# Every report exposes a :meth:`~skore.EstimatorReport.diagnosis` accessor which provides
+# access to several methods:
+# - `.summarize()`` to run checks and get a summary of the findings
+# - `.add()` to add custom checks
+# - `.remove()` to remove checks
+# - `.available()` to list the available checks
+# Let's use `.summarize()` to see what issues can be found for the linear model.
 
 from skore import evaluate
