+++ skore/src/skore/_sklearn/_comparison/__init__.py
@@ -4,7 +4,7 @@ from skore._sklearn._comparison.inspection_accessor import (
 )
 from skore._sklearn._comparison.metrics_accessor import _MetricsAccessor
 from skore._sklearn._comparison.report import ComparisonReport
-from skore._sklearn._diagnostic.accessor import _DiagnosisAccessor
+from skore._sklearn._diagnosis.accessor import _DiagnosisAccessor
 
 _register_accessor("metrics", ComparisonReport)(_MetricsAccessor)
 _register_accessor("inspection", ComparisonReport)(_InspectionAccessor)
