+++ skore/src/skore/_sklearn/_diagnosis/__init__.py
@@ -1,0 +1,4 @@
+from skore._sklearn._diagnosis.base import Check, DiagnosticDisplay
+from skore._sklearn._diagnosis.utils import CheckNotApplicable
+
+__all__ = ["Check", "DiagnosticDisplay", "CheckNotApplicable"]
