+++ skore/src/skore/_sklearn/_diagnosis/base.py
