+++ skore/src/skore/_sklearn/_diagnosis/utils.py
