--- skore/src/skore/_sklearn/_diagnostic/__init__.py
@@ -1,4 +1,0 @@
-from skore._sklearn._diagnostic.base import Check, DiagnosticDisplay
-from skore._sklearn._diagnostic.utils import CheckNotApplicable
-
-__all__ = ["Check", "DiagnosticDisplay", "CheckNotApplicable"]
