+++ skore/tests/unit/reports/cross_validation/test_diagnosis.py
